Switched the Dakota published to Surfpack defaults for the new Kriging model to have a reduced quadratic for the trend function.
Note to Laura, this means that the Paviani 10D 2500 build points 10K eval points case will calculate the negative log likelihood function 241+1 times...compared to the 189+1 evaluations the new kriging was doing with the constant trend function, and the 21+1 evaluations that the Dakota GP does.   So I expect the time the new Kriging model to run this test on your machine will become approximately 469*242/22=5160 seconds... the Dakota GP takes 469 seconds run the test on Laura's machine, and on her machine the new Kriging is getting about the same performance as the Dakota GP.   I haven't yet diagnosed why we're not seeing the same speed up in the new Kriging relative to the Dakota GP on Laura's machine that I am seeing on mine.  TODO...



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@56 aa0779c3-6ec2-4b84-a212-a405c134cd8d