Codifying README.latex into the Makefile.  The following targets now exist for developer convenience when configured --enable-docs:
  doxygen-dev-ref: run doxygen to create dev and ref manuals; invoke latex_modify.perl
  pdf: make pdf versions of dev and ref manual (and via recursion user manual)
  SAND-shell: use SAND-shell-*.tex to make printable SAND reports

Each of these does have a dep on the previous, so you can just 'make SAND-shell' from a clean tree and all three will be run (and a helper target texmf.cnf which is hard coded for RHEL5 that I needed).

The dependencies are tricky to manage due to cyclic dev/ref dependencies, so invoking the PDF targets will always rebuild everything.  I don't expect that to me a common use case, so am not devoting further time to it.

This is intended to be useful, not robust.



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@114 aa0779c3-6ec2-4b84-a212-a405c134cd8d