A few minor updates to COLIN wrapper.  For all practical purposes, this is done.  I'm going to go through and add comments.
Also blessing diffs and updating baselines for all but
coliny_pattern_search in dakota_coliny.in, dakota_coliny_p.in, and
dakota_rosenbrock_users.in.  Carrying pattern search diffs until we
redo blocking/nonblocking and synchronize_nowait(), though I did
replace the parallel failures we've been carrying.  Cobyla diffs were
only in number of function evaluations because function and
constraint evaluations are now combined.  EA diffs were due to
slightly different ordering of a few points in the constrained case,
and the fact that the best point is now guaranteed to be returned with
the final population.

Leaving dakota_hybrid.in, dakota_multiobj3.in, and dakota_scaling.in
up to Laura and Brian.  Only a couple of minor diffs showed up when I
looked at them.




git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@206 aa0779c3-6ec2-4b84-a212-a405c134cd8d