Additions and tweaks for building Dakota within Trilinos using CMake
With Dakota's new CMake infrastructure it is possible to build Dakota from
within Trilinos, where Dakota basically appears as a subdirectory within
TriKota.  To support this, I added some logic to only do the search for
Trilinos if not being built within Trilinos (otherwise the paths to Teuchos
are already set) and added some cache variables to communicate Dakota's
include paths and link libraries back to Trilinos.  There is still a lot
more that could be done with this, leveraging the configure tests and TPL
checks that Trilinos has already done, but this at least gets Dakota building
inside Trilinos with CMake.

Most of the changes in the packages (except pecos) were to install the package
libraries.  The most significant changes to Dakota's and Pecos' CMakeLists
files were to collect all of the subpackage dependencies to send back up to
Trilinos.  This was just kind of hacked in here and could be (should be)
improved.


git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@328 aa0779c3-6ec2-4b84-a212-a405c134cd8d