Change const char* members in ParallelLibrary to std::string.  This was causing problems in library mode if the const char* fell out of scope between calling specify_outputs_restart and manage_outputs_restart.
A positive side effect is we don't have to broadcast dummy strings anymore.  Another I think is we avoid a potential odd pitfall from the old code where a user specifying "" as a filename might get unexpected behavior.

Unforunately the possibility of CommandLineHandler returning a NULL pointer vs. an empty string makes this messier than I'd like (extra conditionals), but that'll eventually get refactored.  The benefit of the std::string outweighs the ugliness.

Also rename std*Flag to std*ToFile to indicate redirection of the stream to file.


git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@382 aa0779c3-6ec2-4b84-a212-a405c134cd8d