Initial cut at CMake system for documentation.  For now, this only works on Unix and is crude, but I need to get feedback from others.
The dependencies are stated correctly in that any of the make targets should do the right thing.  However, the targets are almost all always considered out of date, so for instance 'make docs-doxygen' will run the doxygen phase, and 'make docs-pdf-ref' will needlessly re-run doxygen.

The major improvement that's needed is update User's and Theory builds to use configure_file or relative references so we have dependency to the source directory.

Helpful make targets:

# Make default docs (HTML Dev/Ref, PDF Theory/Users)
... docs
# Run doxygen phase only
... docs-doxygen
# Build all PDF elements, or select ones 
... docs-pdf
... docs-pdf-dev
... docs-pdf-ref
... docs-pdf-theory
... docs-pdf-user
# Build all SAND reports, or select ones
... docs-sand
... docs-sand-dev
... docs-sand-ref
... docs-sand-theory
... docs-sand-user



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@404 aa0779c3-6ec2-4b84-a212-a405c134cd8d