This commit represents a working version of DAKOTA-QUESO integration. I am currently linking against QUESO 0.43. This commit has all of the steps required to instantiate a QUESO environment, set priors on the parameters, calculate a likelihood function, set a proposal density for the MCMC necessary to calculate the posterior parameter values, and actually run the chain and get results. This is a work in progress, and the following issues need to be addressed more comprehensively:
1.  Likelihood function.  Currently the likelihood function just 
involves the residual sum of squares with Gaussian form of the likelihood. 
At a minimum, I want to add an option to allow one to estimate 
the "sigma" error term in the likelihood, assuming iid errors.  
Further extensions to allow several types of likelihood will be in the future.

2.  Data.  Currently we are reading both the "y values" (observed data 
points) and "y std" (the standard deviation at each data point).  
This is consistent with Brian William's GPMSA data file requirements, 
but long term we probably want to condense this with the nonlinear 
least squares data file. 

3.  Emulators.  Currently there is no emulation, but I will add 
a specification in the input file and Mike will finish the NonDBayes class
to allow for GP and stochastic expansion emulators. 

4.  MPI.  ** Currently, if you run QUESO, you must run dakota 
with mpiexec to properly initialize the MPI environment.** 
We will further develop the MPI interface to QUESO over the next few months. 

5. Configuration options.  In this first draft, one needs to have 
the QUESO library installed on their machine.  Bill will change the 
Makefile.include in Dakota/src to reflect the correct settings 
for QUESO.  I do not anticipate any build errors, since I am the only one 
building with QUESO at this point, but please let me know if you have 
problems. The longer term plan is to work with the QUESO team to get some 
type of configuration script/information from QUESO which can be fed to 
the DAKOTA build options. 

6.  QUESO settings.  QUESO has hundreds of knobs which control 
the Markov Chain Monte Carlo algorithms.  We are using many defaults, 
and will be exposing a few to the user (e.g. size of the chain).  As 
I get more familiar with the algorithms, the settings we are using 
may change a bit. 

 


git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@411 aa0779c3-6ec2-4b84-a212-a405c134cd8d