Updates to NIDR and DAKOTA CMake for specificaiton maintenance mode.  Developers should now be able to set ENABLE_SPEC_MAINT and modify dakota.input.nspec and remake.
NIDR will
 * build libnidr always
 * build nidrgen and Dt if ENABLE_SPEC_MAINT
 * rebuild NIDR source files with lex and yacc only if requested

DAKOTA will
 * build dakota.input.summary, dakota.input.desc, NIDR_keywds.H, NIDR_guikeywds.h if ENABLE_SPEC_MAINT
 * TODO: build NIDR_guikeywds.h, dakota.input.desc, dakreord, dakreorder conditionally on target jaguar-files

NOTE: I'm generating these files into the source tree to mimic the autotools build for now and since currently they need to be checked in.  Once we are confident that we can build nidrgen and Dt on all platforms, including MSVS, we can relax this requirement, not check them into the repo, and generate them during the build (well really only the NIDR_keywds.h that's needed).



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@514 aa0779c3-6ec2-4b84-a212-a405c134cd8d