Incremental modifications to Perl script to address Public Ticket #4039 and permit easier in binary tree testing for CMake.  Offers input and output dirs as well as specified file to extract to. New long options:
--extract-to=string  # for intermediate test file name
--input-dir=/path/to/testfilesandbaselines # input source files
--output-dir=/path/to/dest # for any generated intermediate, diff, and baseline files

Any long option in Perl can be abbreviated to the shortest unique and in many cases the equal sign can be omitted, so "-i /path/to" will suffice in this example. 

Also tightened up the usage documentation for serial vs. parallel cases and included information on specifying multiple test files or globs.




git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@584 aa0779c3-6ec2-4b84-a212-a405c134cd8d