Moved get_cwd() from WorkdirHelper to filesystem_utils since it more logically belongs there (and has dependencies on dakdrive and Map() for Windows builds).
Eliminated return value from WorkdirHelper::get_dakpath() since the returned
object does not provide any added value now that get_cwd() can be called
when needed.

Changed the logic for calling WorkdirHelper::get_dakpath() in the 3 "clients":
  not_executable(), Symlink(), and get_npath().

Prior to this commit WorkdirHelper::get_dakpath was not called after the cwdBegin
pointer was set to non-NULL.  Now, get_dakpath() is called unconditionally since
it seems safe to do so AND seems (to me) to be the intended behavior.



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@614 aa0779c3-6ec2-4b84-a212-a405c134cd8d