Added a new helper function, putenv_impl(const std::string& path), based on some concepts that I learned while browsing Boost's testing header files (NOT provided in any of the Boost libraries).
It encapsulates the SunPro_CC stupidity in a single location and provides
an easy transition path from putenv to setenv should the latter be deemed
the more modern/portable way to modify the user's $PATH environment variable.

(Not sure that modifying the user's $PATH is a good idea, but it is what it
is.  Anyone with comments/concerns/enlightenment regarding modification of
the user's environment please send email or call!)



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@621 aa0779c3-6ec2-4b84-a212-a405c134cd8d