Added code to address cases 2, 4, and 5 in Trac ticket #4051.
Added more conditional compilation cases to the
WorkHelper::set_preferred_env_path() method for platform portability
when determining the directory where dakota executable lives.

Added a new method to WorkdirHelper
  ::prepend_preferred_env_path(extra_path)
and called the method from NIDR so that the templatedir is now in the $PATH.

Request review from Brian Adams.



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@632 aa0779c3-6ec2-4b84-a212-a405c134cd8d