Re-enabled key functionality from commit r632: https://software.sandia.gov/trac/dakota/changeset/632
In particular, templatedir is used to augment the $PATH.
Verified that all of the tests in Brian's workdir test suite pass.

Request review from Brian, e.g.  do we really want to determine the absolute
path to the dakota executable and use it to augment the $PATH?  (the resulting
logic requires the use of condtional compilation to achieve portability.)




git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@646 aa0779c3-6ec2-4b84-a212-a405c134cd8d