Data fit surrogates: fix bug (reported by Laura) in sequencing of variables/response lists when a surrogate includes data from a file.  Use of negative eval counters for file read meant variables were getting inserted in forward order, but responses in reverse order.
Need to review this approach with Mike, but for now am adding all the PRP reused points in a single function call, then iterating over file reuse points, inserting them one at a time if appropriate for reuse.  This may cause an extra deep copy of Response in make_pair, which we could mitigate if problematic.

Also catch additional exception and remove needless file close in tabular I/O helpers as the objects will be closed as part of destruction when they fall out of scope.


git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@668 aa0779c3-6ec2-4b84-a212-a405c134cd8d