Updates for NonDCalibration management of state variables which come from experimental configuration data.
Update NonDCalibration constructor to manage configuration variables and reconcile them against the potential state variables.  Mike correctly offers that the approach is fragile and a maintenance problem, but I'll need to consult with him on what enhancements to SharedVariablesData will make this sustainable.

Add a helper function for setting state variables on a model; for use in QUESO and GPMSA.  Uses private variables initialized in the constructor for tracking state variable positions in the arrays.

Move experimental data variables to protected scope.



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@697 aa0779c3-6ec2-4b84-a212-a405c134cd8d