Add an option for lightweight-constructed DataFitSurrModel to read data points from a file.  This will enable EGO, EGRA, etc. to build initial surrogates from file data.  This is joint work with Laura.
Additional work will be needed to manage whether the DACE Iterator correctly augments the points when needed and appropriately warns the user.  

Also, when setting up the DACE Iterators in the calling methods, I think it will be necessary to set the number of samples to 0 to avoid having the iterator augment beyond the minimum required, but I haven't traced that call stack yet.



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@949 aa0779c3-6ec2-4b84-a212-a405c134cd8d