Removed "driver" ctest build script.
This script was originally useful in providing a starting point for jenkins-
driven builds/tests but it reached end-of-lifecycle since Dena's scripts/
fragments are better.  Therefore it has been removed to eliminate potential
confusion over which script to launch.

Let me know if anyone is using this script and I'll suggest an alternative
approach to get a DAKOTA build launched from a jenkins job.



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1007 aa0779c3-6ec2-4b84-a212-a405c134cd8d