Changes to the input spec to add the adaptive sampling approaches developed by Dan Maljovec this summer.  Some of the adaptive sampling approaches are based on topological metrics and the Morse-Smale complex, others are based on metrics such as distance or ALM (which picks the point that has maximum predictive variance according to the GP).  This capability is a nice addition to what we currently have.
This is an initial spec change just to get things working.  Currently many 
of the options for the adaptive sampling class are specified under 
misc_options.  We need to expose these more generally, but we need a little 
more thought about the terminology we should use across all adaptive 
methods. 



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@1096 aa0779c3-6ec2-4b84-a212-a405c134cd8d