Changes to the importance sampling specification, with the addition of a separate specification for the Gaussian Process Adaptive Importance Sampling (GPAIS).  GPAIS is working. Currently the controls are max_iterations (this specifies the number of adaptive samples added), emulator_samples (this specifies the number of candidate samples on the GP emulator), and samples (this specifies the number of initial samples used to build the GP).  We may need to change the spec to be consistent with other uses of samples and/or add more keywords.
The "plain" importance sampling based on the variations we use in the reliability methods 
(IS, AIS, MMAIS) is now reactivated and specified with "importance_sampling." 
This method initially generates an LHS sample then uses one of the approaches listed above. 
MMAIS appears to be the most robust.  This still needs debugging, but the basic architecture is 
working. 




git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@1175 aa0779c3-6ec2-4b84-a212-a405c134cd8d