Major changes to the NonDGPMSABayesCalibration class.  This class now interfaces with the GPMSA implemented in QUESO.  The class follows the current API:  it will change a bit, but the basic implementation is working.  Note that the DAKOTA implementation does read the information provided in the DAKOTA input file and instantiates the simulationStorage, simulationModel, experimentalStorage, and experimentModel objects properly.  The class does currently have some hardwired settings to replicate the Tower example. This will be removed when we have the full implementation working. At this point, this DAKOTA-GPMSA implementation is getting the same results as the standalone GPMSA implementation.
TODO:  Update the implemention when we get a new code version in January. 
Determine how DAKOTA will handle functional data. 
Extend our current handling of experimental data to allow for cases 
where the data does not correspond to the same points as the model output. 
Determine better handling/file naming of the output files produced 
by GPMSA. 

This commit also has updates to the QUESO class.  Most notably, 
there were some underlying QUESO class option changes between 45.2 
and 45.3, so this commit has the necessary changes to the new QUESO 
interface. 

Finally, I updated the CMakeLists file with a temporary fix to point 
to the proper dependencies of QUESO and GPMSA, but this needs to be revamped 
when we get GPMSA in more of a production mode. 



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1271 aa0779c3-6ec2-4b84-a212-a405c134cd8d