Blessing diffs in dakota_hybrid.in and dakota_coliny_p.in and updating baseline.  These were all caused by switch to COLIN cache and corresponding change in duplicate evaluation management.
Blessing diff in dakota_su_mogatest1.in and updating baseline.  This
was partly caused by change to convergence operator in JEGA.  Also
partly caused by something in changeset 887 or 888 that affected the
dakota GP responses.  Laura is looking into the latter in the context
of other tests and recommended blessing this one since the Pareto
front looks OK.




git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1447 aa0779c3-6ec2-4b84-a212-a405c134cd8d