Revise Dakota build system to allow build of Dakota + QUESO under Trilinos.  This is complicated since queso and grvy are built as external packages AND TriBITS was taking any entry in TPL_LIBS and assuming it could export it.  For now, settled on just making sure queso and grvy libraries weren't in TPL_LIBS and that Dakota's link directories were propagated to TriKota.  This will also require changes in projects like Albany that need to link executables using Dakota.
Longer rationale/notes for the record:

Issue when TriBITS is building Dakota as part of Trilinos.  In the Trilinos build, any library target that is referenced in the build of the project gets exported to an exported targets file.  Specifically I believe that anything in Dakota_LIBRARIES is assumed part of the build and is exported, and anything in Dakota_EXTRA_LIBRARIES is only considered a deplib and not exported.  (See TriKota/src/CMakeLists.txt:212).

Because QUESO and GRVY are not built as part of the project, including them in Dakota_LIBRARIES causes them to break at export.

However, including them in Dakota_EXTRA_LIBRARIES breaks because TriKota isn’t aware of Dakota’s link_directories (via -L).

I tried using add_library(queso UNKNOWN IMPORTED) to create a target for the import library, but this also doesn’t work, due to Trilinos handling of the library list and we still get a link failure.

Moreover, TriKota needs to know to link against GSL and where to find it.  So link dirs need to be propagated regardless.

The ultimate conclusion I came to is that we should probably manage this by:
 * Adding granularity in Dakota/src for TPL_LIBS (built in Dakota), EXTPROJ_TPL_LIBS (built via external project), and EXTRA_TPL_LIBS (system or user-installed TPLs).  Via this mechanisn, don’t include external project libraries in Dakota_LIBRARIES, only in Dakota_EXTRA_LIBRARIES.  For that matter could just use TPL_LIBS and EXTRA_TPL_LIBS, if it’s okay for queso, grvy to be in the latter.
 * No longer do CheckPackage(QUESO) because that appends TPL_LIBS.
 * Don’t put queso_ext and grvy_ext in EXPORT_TARGETS because they don’t define libraries or binaries.
 * To work around link errors I considered just passing the entire Dakota/src LINK_DIRECTORIES property to Trilinos, but another option would be to have the same granularity w.r.t. link directories.  For now I propose passing Dakota_LINK_DIRECTORIES to TriKota for link.  Really should filter out Trilinos when doing this.
 * Consider using add_library(queso IMPORTED), to be able to refer to it as a target instead of just as a library, but that’s not a high priority.  There’s some chatter on the Cmake mailing list about pros/cons of this.





git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1558 aa0779c3-6ec2-4b84-a212-a405c134cd8d