Add a warning when Dakota compiled without X graphics, so users know to remove the graphics keyword (or at least don't expect graphical output).
Unfortunately this doesn't address the issue we see on Windows where users get the "Can't open display error" because Dakota is built with Cygwin graphics, but the X server then isn't running.  We could probably test this with X calls to try to open the display, but should probably modify the way the motif graphics objects get constructed and trap errors there. 


git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@1675 aa0779c3-6ec2-4b84-a212-a405c134cd8d