Changes to the processing of replicates and experiments within the classes in the Minimizer branch.  I have tested several variations of things and hope this does not affect the nightlies.
The capability to recast twice (e.g. once expand to numExperiments*numReplicates, 
then contract to one objective to use a global optimizer) is not fully 
working.  If you uncomment lines 449 and 453 in DakotaOptimizer, and 
then force the weighting in objective() in DakotaMinimizer on line 1793 
to go through the first case, the twice recast optimization works functionally 
in terms of summing the squared residuals and producing an optimal answer. 
However, the reporting is not correct in that case, and also 
something is not getting initialized properly in the weightings. 




git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1848 aa0779c3-6ec2-4b84-a212-a405c134cd8d