Work with Laura on data and MO transforms.
Allow all objective reduction functions to be modular on number of functions to actually transform.  This facilitates user fns expanding to data, reducing to single objective.  Will likely remove the forwarders and make callers be explicit to reduce errors.

Adjust propagation of weights from submodel for case when they are empty.  Need to do for sense as well, but can probably write a templated helper function.

Next we will work on final results retrieval.



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1875 aa0779c3-6ec2-4b84-a212-a405c134cd8d