Modified to get queso working on ictinus/Darwin. The problem was that libboost_program_options.dylib was not being installed to bin by InstallDarwinDylibs.cmake.
InstallDarwinDylibs.cmake identifies libraries that need to be installed using otool.

Roughly, it first ran otool on the dakota executable, then prepended an absolute path to any boost dylibs it found. As it was doing this, it removed the boost dylibs from the list and added them to a separate list.

Then it looped over the remaining items in the list and ran otool on them to identify any other libraries that needed to be copied. Significantly, if at this stage it found any boost dylibs without absolute paths, it discarded them.

On ictinus, otool -L shows that the queso dylib (libqueso-0.46.dylib) contains a reference to libboost_program_options.dylib which does NOT have an absolute path. Consequently, InstallDarwinDylibs.cmake was discarding this reference, and it wasn't being installed.

The modification I made to InstallDarwinDylibs.cmake adds a second pass over the list of discovered libraries to prepend absolute paths to boost dylibs. It also removes the logic to discard references to boost dylibs that lack absolute paths.

This approach has been empirically validated to "fix" queso on icitinus and to not break it on my macbook, but since I don't understand the purpose of the logic as it was written, this change should definitely be reviewed by whoever originally wrote the file.



git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@1916 aa0779c3-6ec2-4b84-a212-a405c134cd8d