Changes in the baseline for three EGO-based tests. The first one, dakota_cantilever_ego.in, is a DIFF that we have kept for two years. It relates to the dakota GP. It started diffing around Dec. 17, 2011, when Mike made some changes to the data_io.h wrappers. We never resolved it, but it is probably time to change it. I ran this test with varying seeds, and got a range of results around 8 as the objective function.
The other two diffs relate to the change in the default maxIterations for EGO. 
These two diffs (dakota_separable_5D.in and dakota_users_herbie_shubert_opt_ego.in), 
started diffing three days ago and are fixed now.  

----------------------------------------------------------------------------------
Here is a chain of emails from the last release about the cantilever_ego DIFF, just to maintain the issue: 
Email 1/25/2013

Laura,

Do you see the connection between I/O of the Teuchos matrices and the numerical operations of the GP?  I could perhaps see it if we were talking about parallel runs with matrices sent over MPI buffers, but otherwise, I'm confused as to how this could have an effect (?).

Thanks,
--Mike

On 01/25/2013 09:25 AM, Swiler, Laura Painton wrote:
Patty,
The dakota_su_mogatest1 diff (at least the 12/17/2011 diff) is due to a change in the behavior of the GaussProcApproximation class (the gaussian_process dakota specification class).
We have only a few tests covering that, but the other ones (dakota_cantilever_ego and dakota_uq_rosenbrock_users) also diffed that day.  I am chasing down that diff:  it appears to be from that commit you mention (changeset 888), but the reason is probably from the change Mike made in the data_io.h wrappers.  I need to look at this more, but he replaced some binary/MPI stream templates with generic stream templates in data_io.h, specifically the templates for reading and writing the Teuchos serial symmetric dense matrix.  Since we use that in the GaussProcApproximation class extensively for the covariance matrix, I suspect that is the cause.

I have done several tests with the Dakota GP before and after the change.  I do see small numerical differences, but they do not appear to be systematically biased one way or the other:  sometimes the resulting GP is more accurate, sometimes less.


-----------------------------------------------------------------------------------




git-svn-id: https://software.sandia.gov/svn/dakota/core/trunk@1935 aa0779c3-6ec2-4b84-a212-a405c134cd8d