Updates to explain the num_experiments and num_replicates in the experimental data when read from a calibration_data_file.
Brian:  see what you think of this.  There is currently nothing in the User's manual 
about this.  I can add something or leave this discussion entirely in the 
Reference manual.  Also, for now, we are not discussing configuration 
variables...not sure what we should do about those at this point. 



git-svn-id: https://software.sandia.gov/svn/public/dakota/core/trunk@1943 aa0779c3-6ec2-4b84-a212-a405c134cd8d