Workaround for dakota_lib_mode 2 and 3, where callbacks don't work on letters.  For now, we make the callback from the envelope.  Thanks for Mike for the suggestion, though I retain the ticket to further fix this.
I staged a commit where I added forwards to letter in all of the get/set functions, but couldn't get it to work.

Request Mike take a look at parallel pdakota_lib_mode 4,5 in release and 4,5,6,7 in debug, which are failing by running multiple copies of the iterator when explcit server scheduling is enabled.



git-svn-id: https://software.sandia.gov/svn/dakota/trunk@2263 aa0779c3-6ec2-4b84-a212-a405c134cd8d